Organize code into dir and classes
