+++ decorators/base_decorator.rb
@@ -1,0 +1,13 @@
+require_relative 'interfaces/nameable'
+
+# Base Decorator
+class BaseDecorator < Nameable
+  def initialize(nameable)
+    super()
+    @nameable = nameable
+  end
+
+  def correct_name
+    @nameable.correct_name
+  end
+end
