+++ decorators/capitalize_decorator.rb
@@ -1,0 +1,8 @@
+require_relative 'base_decorator'
+
+# CapitalizeDecorator
+class CapitalizeDecorator < BaseDecorator
+  def correct_name
+    super().to_s.capitalize
+  end
+end
