+++ decorators/trimmer_decorator.rb
@@ -1,0 +1,8 @@
+require_relative 'base_decorator'
+
+# TrimmerDecorator
+class TrimmerDecorator < BaseDecorator
+  def correct_name
+    super()[0..9]
+  end
+end
