+++ lib/person.rb
@@ -1,5 +1,4 @@
-require_relative 'nameable'
-require_relative 'base_decorator'
+require_relative 'interfaces/nameable'
 
 # Person class inheriting from Nameable
 class Person < Nameable
@@ -28,10 +27,3 @@ class Person < Nameable
     @age >= 18
   end
 end
-
-# person = Person.new('Maximilianus', 22)
-# person.correct_name
-# capitalized_person = CapitalizeDecorator.new(person)
-# puts capitalized_person.correct_name
-# capitalized_trimmed_person = TrimmerDecorator.new(capitalized_person)
-# puts capitalized_trimmed_person.correct_name
