+++ lib/student.rb
